+++ bindings/matrix-sdk-crypto-js/tests/device.test.js
@@ -171,6 +171,9 @@ describe("Key Verification", () => {
         // Verification request for `m1`.
         let verificationRequest1;
 
+        /** registerChangesCallback function for `verificationRequest1` */
+        const verificationRequest1ChangesCallback = jest.fn().mockImplementation(() => Promise.resolve());
+
         // The flow ID.
         let flowId;
 
@@ -213,6 +216,9 @@ describe("Key Verification", () => {
 
             expect(outgoingVerificationRequest).toBeInstanceOf(ToDeviceRequest);
             expect(outgoingVerificationRequest.event_type).toStrictEqual("m.key.verification.request");
+
+            // add an onChange listener so we can check it is called at the right times
+            verificationRequest1.registerChangesCallback(verificationRequest1ChangesCallback);
 
             const toDeviceEvents = [
                 {
@@ -282,11 +288,15 @@ describe("Key Verification", () => {
                 },
             ];
 
+            expect(verificationRequest1ChangesCallback).not.toHaveBeenCalled();
+
             // Let's send the verification ready to `m1`.
             await m1.receiveSyncChanges(JSON.stringify(toDeviceEvents), new DeviceLists(), new Map(), new Set());
         });
 
         test("verification requests are synchronized and automatically updated", () => {
+            // receiving the update should have caused a call to the callback
+            expect(verificationRequest1ChangesCallback).toHaveBeenCalledTimes(1);
             expect(verificationRequest1.isReady()).toStrictEqual(true);
             expect(verificationRequest2.isReady()).toStrictEqual(true);
             expect(verificationRequest1.phase()).toStrictEqual(VerificationRequestPhase.Ready);
@@ -357,6 +367,9 @@ describe("Key Verification", () => {
         // SAS verification for the second machine.
         let sas1;
 
+        /** registerChangesCallback function for `sas1` */
+        const sas1ChangesCallback = jest.fn().mockImplementation(() => Promise.resolve());
+
         test("can fetch and accept an ongoing SAS verification (`m.key.verification.accept`)", async () => {
             expect(verificationRequest1.phase()).toStrictEqual(VerificationRequestPhase.Transitioned);
 
@@ -390,7 +403,10 @@ describe("Key Verification", () => {
             expect(sas1Again).toBeInstanceOf(Sas);
             expect(sas1Again.flowId).toStrictEqual(flowId);
 
-            // Let's accept thet SAS start request.
+            // add an onChange listener so we can check it is called at the right times
+            sas1.registerChangesCallback(sas1ChangesCallback);
+
+            // Let's accept that SAS start request.
             let outgoingVerificationRequest = sas1.accept();
 
             expect(outgoingVerificationRequest).toBeInstanceOf(ToDeviceRequest);
@@ -406,8 +422,15 @@ describe("Key Verification", () => {
                 },
             ];
 
+            // The changes callback should not yet have been called
+            expect(sas1ChangesCallback).not.toHaveBeenCalled();
+
             // Let's send the SAS accept to `m2`.
             await m2.receiveSyncChanges(JSON.stringify(toDeviceEvents), new DeviceLists(), new Map(), new Set());
+
+            // ... which should trigger the changes callback
+            expect(sas1ChangesCallback).toHaveBeenCalledTimes(1);
+            sas1ChangesCallback.mockClear();
         });
 
         test("emojis are supported by both sides", () => {
@@ -431,10 +454,12 @@ describe("Key Verification", () => {
                 },
             ];
 
-            // Let's send te SAS key to `m1`.
+            // Let's send the SAS key to `m1`.
             await m1.receiveSyncChanges(JSON.stringify(toDeviceEvents), new DeviceLists(), new Map(), new Set());
-
             m2.markRequestAsSent(toDeviceRequest.id, toDeviceRequest.type, "{}");
+
+            // ... which should trigger the changes callback
+            expect(sas1ChangesCallback).toHaveBeenCalledTimes(1);
         });
 
         test("other side sends back verification key (`m.key.verification.key`)", async () => {
@@ -784,6 +809,9 @@ describe("Key Verification", () => {
         // QR verification for the second machine.
         let qr2;
 
+        /** registerChangesCallback function for `qr1` */
+        const qr2ChangesCallback = jest.fn().mockImplementation(() => Promise.resolve());
+
         test("can generate a QR code", async () => {
             qr2 = await verificationRequest2.generateQrCode();
 
@@ -802,6 +830,8 @@ describe("Key Verification", () => {
             expect(qr2.reciprocated()).toStrictEqual(false);
             expect(qr2.flowId).toMatch(/^[a-f0-9]+$/);
             expect(qr2.roomId).toBeUndefined();
+
+            qr2.registerChangesCallback(qr2ChangesCallback);
         });
 
         test("can read QR code's bytes", async () => {
@@ -929,8 +959,14 @@ describe("Key Verification", () => {
                 },
             ];
 
+            // qr2's changes callback should not yet have been called
+            expect(qr2ChangesCallback).not.toHaveBeenCalled();
+
             // Let's send the verification request to `m2`.
             await m2.receiveSyncChanges(JSON.stringify(toDeviceEvents), new DeviceLists(), new Map(), new Set());
+
+            // ... which should trigger its changes callback
+            expect(qr2ChangesCallback).toHaveBeenCalledTimes(1);
         });
 
         test("can confirm QR code has been scanned", () => {