+++ crates/matrix-sdk-ui/tests/integration/room_list.rs
@@ -8,7 +8,7 @@ use matrix_sdk_test::async_test;
 use matrix_sdk_ui::{
     room_list::{
         EntriesLoadingState, Error, Input, RoomListEntry, State, ALL_ROOMS_LIST_NAME as ALL_ROOMS,
-        VISIBLE_ROOMS_LIST_NAME as VISIBLE_ROOMS,
+        INVITES_LIST_NAME as INVITES, VISIBLE_ROOMS_LIST_NAME as VISIBLE_ROOMS,
     },
     timeline::{TimelineItem, VirtualTimelineItem},
     RoomList,
@@ -231,6 +231,11 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                         "is_tombstoned": false,
                         "not_room_types": ["m.space"],
                     },
+                    "bump_event_types": [
+                        "m.room.message",
+                        "m.room.encrypted",
+                        "m.sticker",
+                    ],
                     "sort": ["by_recency", "by_name"],
                     "timeline_limit": 1,
                 },
@@ -287,6 +292,22 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                     "sort": ["by_recency", "by_name"],
                     "timeline_limit": 20,
                 },
+                INVITES: {
+                    "ranges": [[0, 99]],
+                    "required_state": [
+                        ["m.room.avatar", ""],
+                        ["m.room.encryption", ""],
+                        ["m.room.member", "$ME"],
+                        ["m.room.canonical_alias", ""],
+                    ],
+                    "filters": {
+                        "is_invite": true,
+                        "is_tombstoned": false,
+                        "not_room_types": ["m.space"],
+                    },
+                    "sort": ["by_recency", "by_name"],
+                    "timeline_limit": 0,
+                },
             },
         },
         respond with = {
@@ -300,6 +321,10 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                 },
                 VISIBLE_ROOMS: {
                     "count": 0,
+                    "ops": [],
+                },
+                INVITES: {
+                    "count": 2,
                     "ops": [],
                 },
             },
@@ -327,6 +352,9 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     "ranges": [[0, 19]],
                 },
+                INVITES: {
+                    "ranges": [[0, 1]],
+                }
             },
         },
         respond with = {
@@ -340,6 +368,10 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                 },
                 VISIBLE_ROOMS: {
                     "count": 0,
+                    "ops": [],
+                },
+                INVITES: {
+                    "count": 3,
                     "ops": [],
                 },
             },
@@ -362,6 +394,9 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     "ranges": [[0, 19]],
                 },
+                INVITES: {
+                    "ranges": [[0, 2]],
+                },
             },
         },
         respond with = {
@@ -377,6 +412,10 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                     "count": 0,
                     "ops": [],
                 },
+                INVITES: {
+                    "count": 0,
+                    "ops": [],
+                }
             },
             "rooms": {
                 // let's ignore them for now
@@ -395,7 +434,10 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                     "ranges": [[0, 199]],
                 },
                 VISIBLE_ROOMS: {
-                    "ranges": [],
+                    "ranges": [[0, 19]],
+                },
+                INVITES: {
+                    "ranges": [[0, 0]],
                 },
             },
         },
@@ -409,6 +451,10 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                     ],
                 },
                 VISIBLE_ROOMS: {
+                    "count": 0,
+                    "ops": [],
+                },
+                INVITES: {
                     "count": 0,
                     "ops": [],
                 },
@@ -476,6 +522,9 @@ async fn test_sync_resumes_from_previous_state() -> Result<(), Error> {
                     VISIBLE_ROOMS: {
                         "ranges": [[0, 19]],
                     },
+                    INVITES: {
+                        "ranges": [[0, 99]],
+                    },
                 },
             },
             respond with = {
@@ -486,6 +535,10 @@ async fn test_sync_resumes_from_previous_state() -> Result<(), Error> {
                         "ops": [],
                     },
                     VISIBLE_ROOMS: {
+                        "count": 0,
+                        "ops": [],
+                    },
+                    INVITES: {
                         "count": 0,
                         "ops": [],
                     },
@@ -511,6 +564,9 @@ async fn test_sync_resumes_from_previous_state() -> Result<(), Error> {
                     VISIBLE_ROOMS: {
                         "ranges": [[0, 19]],
                     },
+                    INVITES: {
+                        "ranges": [[0, 0]],
+                    },
                 },
             },
             respond with = {
@@ -521,6 +577,10 @@ async fn test_sync_resumes_from_previous_state() -> Result<(), Error> {
                         "ops": [],
                     },
                     VISIBLE_ROOMS: {
+                        "count": 0,
+                        "ops": [],
+                    },
+                    INVITES: {
                         "count": 0,
                         "ops": [],
                     },
@@ -605,6 +665,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                     // Hello new list.
                     "ranges": [[0, 19]],
                 },
+                INVITES: {
+                    // Hello new list.
+                    "ranges": [[0, 99]],
+                },
             },
         },
         respond with = (code 400) {
@@ -638,6 +702,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                     // We have set a viewport, which reflects here.
                     "ranges": [[5, 10]],
                 },
+                INVITES: {
+                    // The range hasn't been modified due to previous error.
+                    "ranges": [[0, 99]],
+                },
             },
         },
         respond with = {
@@ -646,6 +714,9 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                 ALL_ROOMS: {
                     "count": 110,
                 },
+                INVITES: {
+                    "count": 3,
+                }
             },
             "rooms": {},
         },
@@ -666,6 +737,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     // Despites the error, the range is kept.
                     "ranges": [[5, 10]],
+                },
+                INVITES: {
+                    // Despites the error, the range has made progress.
+                    "ranges": [[0, 2]],
                 },
             },
         },
@@ -696,6 +771,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                     // Despites the error, the range is kept.
                     "ranges": [[5, 10]],
                 },
+                INVITES: {
+                    // Despites the error, the range is kept.
+                    "ranges": [[0, 2]],
+                }
             },
         },
         respond with = {
@@ -703,6 +782,9 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
             "lists": {
                 ALL_ROOMS: {
                     "count": 110,
+                },
+                INVITES: {
+                    "count": 0,
                 },
             },
             "rooms": {},
@@ -723,6 +805,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     // No error. The range is still here.
                     "ranges": [[5, 10]],
+                },
+                INVITES: {
+                    // The range is making progress.
+                    "ranges": [[0, 0]],
                 },
             },
         },
@@ -752,6 +838,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     // The range is still here.
                     "ranges": [[5, 10]],
+                },
+                INVITES: {
+                    // The range is kept as it was.
+                    "ranges": [[0, 0]],
                 },
             },
         },
@@ -783,6 +873,10 @@ async fn test_sync_resumes_from_terminated() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     // The range is still here.
                     "ranges": [[5, 10]],
+                },
+                INVITES: {
+                    // The range is kept as it was.
+                    "ranges": [[0, 0]],
                 },
             },
         },
@@ -866,7 +960,7 @@ async fn test_entries_stream() -> Result<(), Error> {
         set[1] [ F("!r1:bar.org") ];
         set[2] [ F("!r2:bar.org") ];
         pending;
-    }
+    };
 
     sync_then_assert_request_and_fake_response! {
         [server, room_list, sync]
@@ -874,12 +968,13 @@ async fn test_entries_stream() -> Result<(), Error> {
         assert request >= {
             "lists": {
                 ALL_ROOMS: {
-                    "ranges": [
-                        [0, 9],
-                    ],
+                    "ranges": [[0, 9]],
                 },
                 VISIBLE_ROOMS: {
                     "ranges": [[0, 19]],
+                },
+                INVITES: {
+                    "ranges": [[0, 99]],
                 },
             },
         },
@@ -900,15 +995,15 @@ async fn test_entries_stream() -> Result<(), Error> {
                         {
                             "op": "INSERT",
                             "index": 0,
-                            "room_id": "!r3:bar.org"
+                            "room_id": "!r3:bar.org",
                         },
                     ],
                 },
                 VISIBLE_ROOMS: {
                     "count": 0,
-                    "ops": [
-                        // let's ignore them for now
-                    ],
+                },
+                INVITES: {
+                    "count": 0,
                 },
             },
             "rooms": {
@@ -927,7 +1022,7 @@ async fn test_entries_stream() -> Result<(), Error> {
         remove[0];
         insert[0] [ F("!r3:bar.org") ];
         pending;
-    }
+    };
 
     Ok(())
 }
@@ -1007,6 +1102,9 @@ async fn test_entries_stream_with_updated_filter() -> Result<(), Error> {
                 VISIBLE_ROOMS: {
                     "ranges": [[0, 19]],
                 },
+                INVITES: {
+                    "ranges": [[0, 99]],
+                },
             },
         },
         respond with = {
@@ -1029,9 +1127,9 @@ async fn test_entries_stream_with_updated_filter() -> Result<(), Error> {
                 },
                 VISIBLE_ROOMS: {
                     "count": 0,
-                    "ops": [
-                        // let's ignore them for now
-                    ],
+                },
+                INVITES: {
+                    "count": 0,
                 },
             },
             "rooms": {
@@ -1064,6 +1162,162 @@ async fn test_entries_stream_with_updated_filter() -> Result<(), Error> {
         [entries_stream]
         insert[1] [ F("!r1:bar.org") ];
         insert[2] [ F("!r4:bar.org") ];
+        pending;
+    };
+
+    Ok(())
+}
+
+#[async_test]
+async fn test_invites_stream() -> Result<(), Error> {
+    let (server, room_list) = new_room_list().await?;
+
+    let sync = room_list.sync();
+    pin_mut!(sync);
+
+    // The invites aren't accessible yet.
+    assert!(room_list.invites().await.is_err());
+
+    sync_then_assert_request_and_fake_response! {
+        [server, room_list, sync]
+        states = Init => FirstRooms,
+        assert request = {
+            "lists": {
+                ALL_ROOMS: {
+                    "ranges": [[0, 19]],
+                },
+            },
+        },
+        respond with = {
+            "pos": "0",
+            "lists": {
+                ALL_ROOMS: {
+                    "count": 0,
+                },
+            },
+            "rooms": {},
+        },
+    };
+
+    // The invites aren't accessible yet.
+    assert!(room_list.invites().await.is_err());
+
+    let room_id_0 = room_id!("!r0:bar.org");
+
+    sync_then_assert_request_and_fake_response! {
+        [server, room_list, sync]
+        states = FirstRooms => AllRooms,
+        assert request = {
+            "lists": {
+                ALL_ROOMS: {
+                    "ranges": [[0, 0]],
+                },
+                VISIBLE_ROOMS: {
+                    "ranges": [[0, 19]],
+                },
+                INVITES: {
+                    "ranges": [[0, 99]],
+                },
+            },
+        },
+        respond with = {
+            "pos": "1",
+            "lists": {
+                ALL_ROOMS: {
+                    "count": 0,
+                },
+                VISIBLE_ROOMS: {
+                    "count": 0,
+                },
+                INVITES: {
+                    "count": 1,
+                    "ops": [
+                        {
+                            "op": "SYNC",
+                            "range": [0, 0],
+                            "room_ids": [
+                                room_id_0,
+                            ],
+                        },
+                    ],
+                },
+            },
+            "rooms": {
+                room_id_0: {
+                    "name": "Invitation for Room #0",
+                    "initial": true,
+                },
+            },
+        },
+    };
+
+    let (previous_invites, invites_stream) = room_list.invites().await?;
+    pin_mut!(invites_stream);
+
+    assert_eq!(previous_invites.len(), 1);
+    assert_matches!(&previous_invites[0], RoomListEntry::Filled(room_id) => {
+        assert_eq!(room_id, room_id_0);
+    });
+
+    assert_entries_stream! {
+        [invites_stream]
+        pending;
+    };
+
+    sync_then_assert_request_and_fake_response! {
+        [server, room_list, sync]
+        states = AllRooms => CarryOn,
+        assert request = {
+            "lists": {
+                ALL_ROOMS: {
+                    "ranges": [[0, 0]],
+                },
+                VISIBLE_ROOMS: {
+                    "ranges": [[0, 19]],
+                },
+                INVITES: {
+                    "ranges": [[0, 0]],
+                },
+            },
+        },
+        respond with = {
+            "pos": "2",
+            "lists": {
+                ALL_ROOMS: {
+                    "count": 0,
+                },
+                VISIBLE_ROOMS: {
+                    "count": 0,
+                },
+                INVITES: {
+                    "count": 1,
+                    "ops": [
+                        {
+                            "op": "DELETE",
+                            "index": 0,
+                        },
+                        {
+
+                            "op": "INSERT",
+                            "index": 0,
+                            "room_id": "!r1:bar.org",
+                        },
+                    ],
+                },
+            },
+            "rooms": {
+                "!r1:bar.org": {
+                    "name": "Invitation for Room #1",
+                    "initial": true,
+                },
+            },
+        },
+    };
+
+    assert_entries_stream! {
+        [invites_stream]
+        remove[0];
+        insert[0] [ F("!r1:bar.org") ];
         pending;
     };
 
@@ -1595,6 +1849,9 @@ async fn test_input_viewport() -> Result<(), Error> {
                     "ranges": [[0, 19]],
                     "timeline_limit": 20,
                 },
+                INVITES: {
+                    "ranges": [[0, 99]],
+                },
             },
         },
         respond with = {
@@ -1617,8 +1874,11 @@ async fn test_input_viewport() -> Result<(), Error> {
                 },
                 VISIBLE_ROOMS: {
                     "ranges": [[10, 15], [20, 25]],
-                }
-            }
+                },
+                INVITES: {
+                    "ranges": [[0, 99]],
+                },
+            },
         },
         respond with = {
             "pos": "1",