+++ bindings/matrix-sdk-crypto-js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.1.0-alpha.11
 
+-   Add `VerificationRequest.registerChangesCallback()`,
+    `Sas.registerChangesCallback()`, and `Qr.registerChangesCallback()`.
 -   Add `VerificationRequest.phase()` and `VerificationRequest.getVerification()`.
 
 # v0.1.0-alpha.10